Factor out UploadSamplesHelper
