+++ experiment/src/org/labkey/experiment/api/SampleTypeServiceImpl.java
@@ -91,7 +91,6 @@ import org.labkey.api.study.publish.StudyPublishService;
 import org.labkey.api.util.CPUTimer;
 import org.labkey.api.util.Pair;
 import org.labkey.api.util.StringUtilsLabKey;
-import org.labkey.experiment.samples.UploadSamplesHelper;
 
 import java.sql.SQLException;
 import java.time.LocalDateTime;
@@ -1026,7 +1025,7 @@ public class SampleTypeServiceImpl extends AbstractAuditHandler implements Sampl
     public boolean parentAliasHasCorrectFormat(String parentAlias)
     {
         //check if it is of the expected format or targeting the to be created sample type
-        if (!(UploadSamplesHelper.isInputOutputHeader(parentAlias) || NEW_SAMPLE_TYPE_ALIAS_VALUE.equals(parentAlias)))
+        if (!(ExperimentService.isInputOutputColumn(parentAlias) || NEW_SAMPLE_TYPE_ALIAS_VALUE.equals(parentAlias)))
             throw new IllegalArgumentException(String.format("Invalid parent alias header: %1$s", parentAlias));
 
         return true;
@@ -1071,7 +1070,6 @@ public class SampleTypeServiceImpl extends AbstractAuditHandler implements Sampl
         return createAuditRecord(c, comment, null, row, null);
     }
 
-    // move to UploadSamplesHelper?
     private boolean isInputFieldKey(String fieldKey)
     {
         int slash = fieldKey.indexOf('/');
